# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMPLR API Documentation",
        "description": "Welcome to the CMPLR API Documentation!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "CMPLR API Server"
        }
    ],
    "paths": {
...
        "/blog/{blog-identifier}/info": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Retrieve Blog Info",
                "description": "This method returns general information about the blog, such as the title, number of posts, and other high-level data.",
                "operationId": "getBlogInfo",
                "parameters": [
                    {
                        "name": "blog-identifier",
                        "in": "path",
                        "description": "Any blog identifier",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://www.example.com/image/avatar.png"
                                        },
                                        "posts": {
                                            "type": "number",
                                            "example": 69
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "john-doe"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://www.cmplr.com/blogs/john-doe"
                                        },
                                        "updated": {
                                            "type": "number",
                                            "example": 1308953007
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "<p><strong>Mr. Karp</strong> is tall and skinny, with unflinching blue eyes a mop of brown hair.\\r\\nHe speaks incredibly fast and in complete paragraphs.</p>"
                                        },
                                        "ask": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "ask_anon": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "likes": {
                                            "type": "number",
                                            "example": 420
                                        },
                                        "is_blocked_from_primary": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "theme": {
                                            "properties": {
                                                "avatar_shape": {
                                                    "type": "string",
                                                    "example": "circle"
                                                },
                                                "background_color": {
                                                    "type": "string",
                                                    "example": "#00f980"
                                                },
                                                "body_font": {
                                                    "type": "string",
                                                    "example": "Arial"
                                                },
                                                "header_image": {
                                                    "type": "string",
                                                    "example": "https://www.example.com/image/header.png"
                                                },
                                                "link_color": {
                                                    "type": "string",
                                                    "example": "#eefeef"
                                                },
                                                "show_avatar": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "show_description": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "show_header_image": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "show_title": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "title_color": {
                                                    "type": "string",
                                                    "example": "#f0f0f0"
                                                },
                                                "title_font": {
                                                    "type": "string",
                                                    "example": "Arial"
                                                },
                                                "title_font_weight": {
                                                    "type": "string",
                                                    "example": "bold"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/{blog-identifier}/likes": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Retrieve Blog's Likes",
                "description": "This method can be used to retrieve the publicly exposed likes from a blog.",
                "operationId": "getBlogLikes",
                "parameters": [
                    {
                        "name": "blog-identifier",
                        "in": "path",
                        "description": "Any blog identifier",
                        "required": true
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Block number to start at",
                        "required": false
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The number of blocks to retrieve, 1-20, inclusive",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "liked_posts": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "blog_name": {
                                                        "type": "string",
                                                        "example": "john-doe"
                                                    },
                                                    "blog_url": {
                                                        "type": "string",
                                                        "example": "https://www.cmplr.com/blogs/john-doe"
                                                    },
                                                    "post_id": {
                                                        "type": "number",
                                                        "example": "1234567890"
                                                    },
                                                    "post_notes_count": {
                                                        "type": "number",
                                                        "example": 20
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "liked_count": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "_links": {
                                            "properties": {
                                                "next": {
                                                    "properties": {
                                                        "href": {
                                                            "type": "string",
                                                            "example": "/api/v1/blogs/john-doe/likes?offset=20"
                                                        },
                                                        "method": {
                                                            "type": "string",
                                                            "example": "GET"
                                                        },
                                                        "query_params": {
                                                            "properties": {
                                                                "offset": {
                                                                    "type": "number",
                                                                    "example": 20
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/post/delete": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "delete existing post",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "the ID of the post to delete",
                        "required": true,
                        "schema": {
                            "type": "Number"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Meta": {
                                            "properties": {
                                                "Status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "example": "OK"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}